refactor: 커스텀 예외 패키지를 발생시키는 곳으로 이동
