+++ src/main/java/roomescape/domain/exception/InvalidNameException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.domain.exception;
+
+import roomescape.exception.CustomException;
 
 public class InvalidNameException extends CustomException {
