+++ src/main/java/roomescape/domain/exception/InvalidRequestException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.domain.exception;
+
+import roomescape.exception.CustomException;
 
 public class InvalidRequestException extends CustomException {
