+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,10 +7,10 @@ import roomescape.controller.reservation.ReservationResponse;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.exception.DuplicateReservation;
-import roomescape.exception.PreviousTimeException;
-import roomescape.exception.ThemeNotFoundException;
-import roomescape.exception.TimeNotFoundException;
+import roomescape.service.exception.DuplicateReservation;
+import roomescape.service.exception.PreviousTimeException;
+import roomescape.service.exception.ThemeNotFoundException;
+import roomescape.service.exception.TimeNotFoundException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
