+++ src/main/java/roomescape/service/exception/TimeNotFoundException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.service.exception;
+
+import roomescape.exception.CustomException;
 
 public class TimeNotFoundException extends CustomException {
