+++ src/main/java/roomescape/service/exception/TimeUsedException.java
@@ -1,4 +1,6 @@
-package roomescape.exception;
+package roomescape.service.exception;
+
+import roomescape.exception.CustomException;
 
 public class TimeUsedException extends CustomException {
