+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -10,9 +10,9 @@ import roomescape.controller.reservation.ReservationRequest;
 import roomescape.controller.reservation.ReservationResponse;
 import roomescape.controller.theme.ReservationThemeResponse;
 import roomescape.controller.time.TimeResponse;
-import roomescape.exception.DuplicateReservation;
-import roomescape.exception.PreviousTimeException;
-import roomescape.exception.TimeNotFoundException;
+import roomescape.service.exception.DuplicateReservation;
+import roomescape.service.exception.PreviousTimeException;
+import roomescape.service.exception.TimeNotFoundException;
 import roomescape.repository.H2ReservationRepository;
 import roomescape.repository.H2ReservationTimeRepository;
 import roomescape.repository.H2ThemeRepository;
